Updated OpponentModelling
OpponentModellingBias knows the true distribution and uses it to calculate the optimal move
BUG FIX: Illegal moves were returning non-zero results at the MCS stage and this was fixed
Tested different powers of likelihood multiplier

+++ java/tud/gamecontroller/players/OPAnytimeHyperPlayer/JointMoveTupleComparator.java
@@ -6,11 +6,11 @@ import tud.gamecontroller.term.TermInterface;
 
 import java.util.Comparator;
 
-public class JointMoveTupleComparator<TermType extends TermInterface> implements Comparator<Tuple<Float, JointMoveInterface<TermType>>> {
+public class JointMoveTupleComparator<TermType extends TermInterface> implements Comparator<Tuple<Double, JointMoveInterface<TermType>>> {
 
     // Overriding compare()method of Comparator
     // for descending order of cgpa
-    public int compare(Tuple<Float, JointMoveInterface<TermType>> m1, Tuple<Float, JointMoveInterface<TermType>> m2) {
+    public int compare(Tuple<Double, JointMoveInterface<TermType>> m1, Tuple<Double, JointMoveInterface<TermType>> m2) {
         if (m1.getA() < m2.getA())
             return 1;
         else if (m1.getA() > m2.getA())
